Move scheduled run handler to a job module
