+++ crawler/main.py
@@ -4,9 +4,6 @@ import time
 
 import pymongo
 
-from flask import current_app as app
-
-from crawler import scheduler
 from crawler.constants import (
     COLLECTION_CENTRES,
     COLLECTION_SAMPLES,
@@ -124,14 +121,3 @@ def run(sftp: bool, keep_files: bool, add_to_dart: bool, settings_module: str = 
         logger.info("=" * 80)
     except Exception as e:
         logger.exception(e)
-
-
-def scheduled_run():
-    """Scheduler's job to do a run every 15 minutes."""
-    logger.info("Starting scheduled_run job.")
-
-    with scheduler.app.app_context():
-        use_sftp = app.config["USE_SFTP"]
-        keep_files = app.config["KEEP_FILES"]
-        add_to_dart = app.config["ADD_TO_DART"]
-        run(use_sftp, keep_files, add_to_dart)
